requirements.txt
asgiref==3.7.2
Django==4.2.3
django-bootstrap5==23.3
django-filter==23.2
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.3.0
Pillow==10.0.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
